% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_dir.R
\name{cache_dir}
\alias{cache_dir}
\title{Title}
\usage{
cache_dir(
  subdir = c("raw", "gisco", "corine"),
  year = c("2021"),
  create_subdir = FALSE
)
}
\arguments{
\item{create_subfolders}{}
}
\description{
Title
}
