% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clc_codes}
\alias{clc_codes}
\title{Corine Land Cover codes}
\format{
\subsection{\code{clc_codes}}{

A data frame with 44 rows and 5 columns:
\describe{
\item{CLC}{Corine Land Cover code}
\item{CLC_Label1}{Description level 1}
\item{CLC_Label2}{Description level 2}
\item{CLC_Label3}{Description level 3}
\item{CLC_RGB}{Suggested RBG colour code for plotting}
}
}
}
\usage{
clc_codes
}
\description{
A list of codes and explanations relating to Corine land use data
}
\keyword{datasets}
