% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_codes}
\alias{country_codes}
\title{Country codes and names}
\format{
\subsection{\code{country_codes}}{

A data frame with 42 rows and 6 columns:
\describe{
\item{Code}{Country code}
\item{Country}{Country name in English}
\item{Eurostat}{Is the country included in Eurostat data?}
\item{Native}{Country name in native/local language}
\item{French}{Country name in French}
\item{German}{Country name in German}
}
}
}
\source{
\url{https://ec.europa.eu/eurostat/statistics-explained/index.php?title=Tutorial:Country_codes_and_protocol_order#EU_and_euro_area_aggregates}
}
\usage{
country_codes
}
\description{
A list of country codes and names that are potentially relevant to EU/NUTS
}
\keyword{datasets}
