% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_corine.R
\name{extract_corine}
\alias{extract_corine}
\title{Extract land use data for a specific country}
\usage{
extract_corine(country_code, refresh = FALSE, verbose = 0L, ...)
}
\arguments{
\item{nuts_year}{}
}
\description{
Extract land use data for a specific country
}
