% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_connectedness.R
\name{generate_connectedness}
\alias{generate_connectedness}
\title{Generate pairwise connectedness between patches using a numerical approximation to integrating over distances between areas within two patches}
\usage{
generate_connectedness(
  patches,
  connectedness_fun,
  max_distance = NULL,
  grid_resolution = 50,
  sparse = TRUE,
  centroid_distance = TRUE,
  verbose = 1L
)
}
\arguments{
\item{patches}{a data frame containing polygons and indexes, such as that created by \code{\link{generate_patches}}}

\item{connectedness_fun}{a vectorised function mapping distances to connectedness i.e. a spatial kernal (TODO: may also take a second argument giving the bearing)}

\item{max_distance}{the maximum distance beyond which connectedness is assumed to be negligable - default is to determine this automatically}

\item{grid_resolution}{the resolution of the numerical approximation, in terms of the number of points to evaluate betweeen zero distance and max_distance}

\item{sparse}{should a sparse data frame be returned, or a dense matrix?}

\item{centroid_distance}{should additional information of connectedness based on centroids be returned (ignored if !isTRUE(sparse))}

\item{verbose}{verbosity level (0 is silent, 2 is noisy)}
}
\description{
Generate pairwise connectedness between patches using a numerical approximation to integrating over distances between areas within two patches
}
