% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_neighbours.R
\name{generate_neighbours}
\alias{generate_neighbours}
\title{Generate a sparse matrix of neighbours for patches (i.e. edges in the graph)}
\usage{
generate_neighbours(
  patches,
  calculate_border = TRUE,
  buffer_dist = 0.001 * hex_width
)
}
\arguments{
\item{patches}{an object created by \code{\link{generate_patches}}}

\item{calculate_border}{should the length of the common border to neighbours be calculated?}

\item{buffer_dist}{a tolerance limit defining which borders are considered to be contiguous}
}
\description{
Generate a sparse matrix of neighbours for patches (i.e. edges in the graph)
}
