% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{hexscape_options}
\alias{hexscape_options}
\alias{hexscape_getOption}
\title{Options for the hexscape package}
\usage{
hexscape_options(...)

hexscape_getOption(name)
}
\arguments{
\item{...}{named option(s) to change - for a list of available options, see details below.}

\item{name}{the name of the option to get the current value of - for a  list of available options, see details below.}
}
\value{
The current value of all available HexScaoe options (after applying any changes specified) is returned invisibly as a named list.
}
\description{
Utility function to set/get options for the hexscape package, most notably the local folder used to store intermediate objects.
}
\details{
The following options can be specified:

\itemize{
\item \bold{storage_folder} - the path to a folder to store intermediate objects created by hexscape.
}
}
\seealso{
\code{\link{hexscape}}
}
\keyword{methods}
