% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_cache_dir.R
\name{hs_cache_dir}
\alias{hs_cache_dir}
\title{Title}
\usage{
hs_cache_dir(
  subdir = c("raw", "gisco", "corine"),
  year = c("2021"),
  create_subdir = TRUE
)
}
\arguments{
\item{create_subfolders}{}
}
\description{
Title
}
