% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_data_dir.R
\name{hs_data_dir}
\alias{hs_data_dir}
\title{Locations for \code{hexscape} data}
\usage{
hs_data_dir(
  subdir = c("gisco", "corine"),
  year = c("2018", "2021"),
  create_subdir = TRUE
)
}
\arguments{
\item{year}{Default to 2021.}

\item{subfolder}{Either \code{gisco} or \code{corine}.}

\item{create_subfolders}{}
}
\description{
Locations for \code{hexscape} data
}
\seealso{
\code{\link[=hs_cache_dir]{hs_cache_dir()}}
}
