% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nuts_codes}
\alias{nuts_codes}
\title{NUTS codes}
\format{
\subsection{\code{nuts_codes}}{

A data frame with 2,016 rows and 5 columns:
\describe{
\item{Code}{Country code}
\item{Country}{Country name}
\item{Level}{NUTS level}
\item{NUTS}{NUTS code}
\item{Label}{NUTS label}
}
}
}
\usage{
nuts_codes
}
\description{
A complete list of available NUTS codes (levels 0/1/2/3)
}
\keyword{datasets}
