% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc_as_cols.R
\name{sfc_as_cols}
\alias{sfc_as_cols}
\title{Title}
\source{
\href{https://github.com/r-spatial/sf/issues/231#issuecomment-424929353}{Source}
}
\usage{
sfc_as_cols(x, geometry, names = c("x", "y"))
}
\arguments{
\item{names}{}
}
\value{

}
\description{
Title
}
