% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_nuts_codes.R
\name{all_nuts_codes}
\alias{all_nuts_codes}
\title{Title}
\usage{
all_nuts_codes(level = 0L:4L, pattern = character(0L), year = "2021")
}
\arguments{
\item{year}{}

\item{nuts_code}{optional pattern match}
}
\description{
Title
}
