% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_corine.R
\name{extract_corine}
\alias{extract_corine}
\alias{validate_corine_cache}
\alias{regenerate_corine_cache}
\alias{read_corine}
\title{Extract land use data from raw polygon file corresponding to a specific map area}
\usage{
extract_corine(
  map,
  corine_path,
  type = c("minimal", "reduced", "full"),
  intersection = TRUE,
  max_rows = 0L,
  verbose = 1L
)

validate_corine_cache()

regenerate_corine_cache(verbose = 1L)

read_corine(nuts_code, year = "2021", corine_path, max_rows = 0L, verbose = 1L)
}
\arguments{
\item{type}{one of "minimal" (only CLC and Shape columns; grouped multi-polygon), "reduced" (only CLC and Shape columns but original rows), or "full" (all columns and rows in underlying data).}

\item{max_rows}{target maxinum number of features to retrieve at a time - this represents a trade off between speed and memory footprint. Special values of 0 means extract each CLC individually (minimum memory requirement), and Inf means all features simultaneously (around 30GB available memory required, for type "minimal" or "reduced").}

\item{verbose}{}
}
\description{
Extract land use data from raw polygon file corresponding to a specific map area
}
