% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_corine.R
\name{extract_corine}
\alias{extract_corine}
\alias{validate_corine_cache}
\alias{regenerate_corine_cache}
\title{Extract land use data for a specific country}
\usage{
extract_corine(
  map,
  use_cache = validate_corine_cache(),
  simplify_keep = 0.25,
  verbose = 1L
)

validate_corine_cache()

regenerate_corine_cache(verbose = 1L)
}
\arguments{
\item{verbose}{}
}
\description{
Extract land use data for a specific country
}
