% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_neighbours.R
\name{generate_neighbours}
\alias{generate_neighbours}
\title{Title}
\usage{
generate_neighbours(
  patches,
  calculate_border = TRUE,
  buffer_dist = 0.001 * hex_width
)
}
\arguments{
\item{buffer_dist}{}
}
\value{

}
\description{
Title
}
