% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_patches.R
\name{generate_patches}
\alias{generate_patches}
\title{Title}
\usage{
generate_patches(
  landscape,
  hex_width,
  land_use = NULL,
  min_prop = 0.01,
  simplify_keep = 0.1
)
}
\arguments{
\item{min_prop}{}
}
\description{
Title
}
\examples{

xrange <- c(0, 50)
yrange <- c(0, 50)
corners <- tribble(~x, ~y,
    xrange[1], yrange[1],
    xrange[2], yrange[1],
    xrange[2], yrange[2],
    xrange[1], yrange[2],
    xrange[1], yrange[1]
)
landscape <- sf::st_sfc(sf::st_multipolygon(list(list(as.matrix(corners)))))
## And add our hexagons:
patches <- generate_patches(landscape, hex_width=2)
patches
ggplot(patches, aes(label=Index)) + geom_sf() + geom_sf_text()

}
