% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexScape-package.R
\docType{package}
\name{HexScape-package}
\alias{HexScape}
\alias{HexScape-package}
\title{HexScape: Aggregation of Spatial Data into Hexagonal Patches using Simple Features}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Matt Denwood \email{md@sund.ku.dk}

Authors:
\itemize{
  \item Mossa Merhi \email{mossa@sund.ku.dk}
}

}
\keyword{internal}
