% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexscape-package.R
\docType{package}
\name{hexscape-package}
\alias{hexscape}
\alias{hexscape-package}
\title{hexscape: Aggregation of Spatial Data into Discrete Patches using Simple Features}
\description{
TODO
}
\author{
\strong{Maintainer}: Matt Denwood \email{md@sund.ku.dk}

Authors:
\itemize{
  \item Mossa Merhi \email{mossa@sund.ku.dk}
}

Other contributors:
\itemize{
  \item Carlijn Bogaardt \email{Carlijn.Bogaardt@glasgow.ac.uk} [contributor]
}

}
\keyword{internal}
