% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_cache_dir.R
\name{hs_cache_dir}
\alias{hs_cache_dir}
\title{Locations for \code{hexscape} cached data}
\usage{
hs_cache_dir(
  subdir = c("gisco", "corine"),
  year = c("2021"),
  create_subdir = TRUE
)
}
\arguments{
\item{year}{Default to 2021.}

\item{subfolder}{Either \code{gisco} or \code{corine}.}

\item{create_subfolders}{}
}
\description{
Locations for \code{hexscape} cached data
}
\seealso{
\code{\link[=hs_data_dir]{hs_data_dir()}}
}
