% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_corine.R
\name{load_corine}
\alias{load_corine}
\title{Extract land use data for a specific NUTS code area}
\usage{
load_corine(
  nuts_codes,
  clc_group = 3L,
  simplify_keep = 1,
  union = TRUE,
  use_cache = TRUE,
  verbose = 1L,
  map_year = hs_year("map"),
  corine_year = hs_year("corine")
)
}
\arguments{
\item{simplify_keep}{passed on to \code{\link[rmapshaper]{ms_simplify}}}
}
\description{
Extract land use data for a specific NUTS code area
}
