% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_corine.R
\name{load_corine}
\alias{load_corine}
\alias{cache_all_corine}
\title{Extract land use data for a specific country}
\usage{
cache_all_corine(exclude = character(0), randomise = FALSE, verbose = 1L)

load_corine(
  nuts_code,
  union = FALSE,
  use_cache = validate_corine_cache(),
  verbose = 1L
)
}
\arguments{
\item{simplify_keep}{passed on to \code{\link[rmapshaper]{ms_simplify}}}
}
\description{
Extract land use data for a specific country
}
