% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_corine.R, R/load_corine.R
\name{load_corine}
\alias{load_corine}
\alias{extract_corine}
\alias{cache_all_corine}
\title{Extract land use data for a specific country}
\usage{
load_corine(
  nuts_code,
  clc_key = NULL,
  corine_path = file.path(hexscape_getOption("storage_folder"), "raw_data",
    "u2018_clc2018_v2020_20u1_geoPackage/DATA/U2018_CLC2018_V2020_20u1.gpkg"),
  verbose = 1L,
  ...
)

extract_corine(
  map,
  simplify_keep = 0.2,
  corine_path = file.path(hexscape_getOption("storage_folder"), "raw_data",
    "u2018_clc2018_v2020_20u1_geoPackage/DATA/U2018_CLC2018_V2020_20u1.gpkg"),
  verbose = 1L
)

cache_all_corine(
  corine_path = file.path(hexscape_getOption("storage_folder"), "raw_data",
    "u2018_clc2018_v2020_20u1_geoPackage/DATA/U2018_CLC2018_V2020_20u1.gpkg"),
  nuts_year = 2016,
  verbose = 1L
)

load_corine(
  nuts_code,
  clc_key = NULL,
  corine_path = file.path(hexscape_getOption("storage_folder"), "raw_data",
    "u2018_clc2018_v2020_20u1_geoPackage/DATA/U2018_CLC2018_V2020_20u1.gpkg"),
  verbose = 1L,
  ...
)

extract_corine(
  map,
  simplify_keep = 0.2,
  corine_path = file.path(hexscape_getOption("storage_folder"), "raw_data",
    "u2018_clc2018_v2020_20u1_geoPackage/DATA/U2018_CLC2018_V2020_20u1.gpkg"),
  verbose = 1L
)
}
\arguments{
\item{simplify_keep}{passed on to \code{\link[rmapshaper]{ms_simplify}}}

\item{nuts_year}{}
}
\description{
Extract land use data for a specific country

Extract land use data for a specific country
}
