% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_map.R
\name{load_map}
\alias{load_map}
\title{Extract a map for given country or NUTS area code}
\usage{
load_map(nuts_code, verbose = 1L)
}
\arguments{
\item{verbose}{}
}
\description{
Extract a map for given country or NUTS area code
}
