% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomise_voronoi.R
\name{randomise_voronoi}
\alias{randomise_voronoi}
\title{Title}
\usage{
randomise_voronoi(
  map,
  points,
  randomise_size = 5L,
  sample_size = 10L,
  max_tries = 3L,
  verbose = 1L
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
\examples{
xrange <- c(0, 10)
yrange <- c(0, 10)
corners <- tribble(~x, ~y,
                   xrange[1], yrange[1],
                   xrange[2], yrange[1],
                   xrange[2], yrange[2],
                   xrange[1], yrange[2],
                   xrange[1], yrange[1]
)
library("sf")
landscape <- st_sfc(st_multipolygon(list(list(as.matrix(corners))))) |> st_as_sf()
points <- st_sfc(lapply(1:10, function(x) st_point(runif(2,0,10)))) |> st_as_sf()
random <- randomise_voronoi(landscape, points) |> mutate(Index = 1:n())
ggplot(random) + geom_sf_label(aes(label=Index)) + geom_sf_label(aes(geometry=RandomPoint, label=Index), col="red")


}
