% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomise_voronoi.R
\name{randomise_voronoi}
\alias{randomise_voronoi}
\title{Title}
\usage{
randomise_voronoi(
  map,
  points,
  randomise_size = 5L,
  sample_size = 10L,
  max_tries = 3L,
  verbose = 1L
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
