% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_map.R
\name{read_map}
\alias{read_map}
\title{Internal \code{read_map} function}
\usage{
read_map(country_code, year)
}
\arguments{
\item{country_code}{ISO 3166-1 Country Code, e.g., \code{"DK"}}

\item{year}{Default to 2021.}
}
\description{
Internal \code{read_map} function
}
