% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_storage_folder.R
\name{set_storage_folder}
\alias{set_storage_folder}
\title{Set storage folder for HexScape package}
\usage{
set_storage_folder(folder)
}
\arguments{
\item{folder}{}
}
\description{
Set storage folder for HexScape package
}
